write extra dataset in HDF5 file
With the current version of the task (07 Jan 2020), some important
fields are important, both to reproduce the dots and to keep track
of which dots were presented in which block.

Thus, I have decided to write a new dataset called paramdset in the
HDF5 file, and this new dataset should have as many rows as the 'px'
dataset.

This feature has NOT been tested yet.